+++ esbuild.config.mjs
@@ -111,6 +111,7 @@ async function buildBackground() {
     readFile("modules/local-libraries.js"),
     readFile("modules/import-action-handler.js"),
     readFile("modules/telemetry-action-handler.js"),
+    readFile("modules/update-action-handler.js"),
     readFile("modules/message-router.js"),
     readFile("modules/gm-audio-handler.js"),
     readFile("modules/gm-menu-handler.js"),