+++ tests/message-response-map.test.js
@@ -10,6 +10,7 @@ import {
 } from "../src/background/message-router.ts";
 import { IMPORT_BACKGROUND_ACTIONS } from "../src/background/import-action-handler.ts";
 import { EXECUTION_TELEMETRY_ACTIONS } from "../src/background/telemetry-action-handler.ts";
+import { UPDATE_BACKGROUND_ACTIONS } from "../src/background/update-action-handler.ts";
 import { GM_AUDIO_ACTIONS } from "../src/background/gm-audio-handler.ts";
 import { GM_COOKIE_ACTIONS } from "../src/background/gm-cookie-handler.ts";
 import { GM_MENU_ACTIONS } from "../src/background/gm-menu-handler.ts";
@@ -23,6 +24,7 @@ import { GM_WEBREQUEST_ACTIONS } from "../src/background/gm-webrequest-handler.t
 const ROUTED_BACKGROUND_ACTIONS = [
   ...IMPORT_BACKGROUND_ACTIONS,
   ...EXECUTION_TELEMETRY_ACTIONS,
+  ...UPDATE_BACKGROUND_ACTIONS,
   ...GM_AUDIO_ACTIONS,
   ...GM_COOKIE_ACTIONS,
   ...GM_MENU_ACTIONS,